feat: добавляет обработчик наведения на карточку
